frontend: Renamed higher level components to better reflect their function.
