+++ packages/frontend/src/components/App.tsx
@@ -1,16 +1,35 @@
-import { FC } from 'react'
-import GreetingForm from '~~/components/GreetingForm'
-import Layout from '~~/components/layout/Layout'
-import NetworkSupportChecker from './NetworkSupportChecker'
+import '@mysten/dapp-kit/dist/index.css'
+import '@radix-ui/themes/styles.css'
+import '@suiware/kit/main.css'
+import SuiProvider from '@suiware/kit/SuiProvider'
+import { FC, StrictMode } from 'react'
+import IndexPage from '~~/components/IndexPage'
+import { APP_NAME } from '~~/config/main'
+import { getThemeSettings } from '~~/helpers/theme'
+import useNetworkConfig from '~~/hooks/useNetworkConfig'
+import ThemeProvider from '~~/providers/ThemeProvider'
+import '~~/styles/index.css'
+import { ENetwork } from '~~/types/ENetwork'
+
+const themeSettings = getThemeSettings()
 
 const App: FC = () => {
+  const { networkConfig } = useNetworkConfig()
+
   return (
-    <Layout>
-      <NetworkSupportChecker />
-      <div className="justify-content flex flex-grow flex-col items-center justify-center rounded-md p-3">
-        <GreetingForm />
-      </div>
-    </Layout>
+    <StrictMode>
+      <ThemeProvider>
+        <SuiProvider
+          customNetworkConfig={networkConfig}
+          defaultNetwork={ENetwork.LOCALNET}
+          walletAutoConnect={false}
+          walletStashedName={APP_NAME}
+          themeSettings={themeSettings}
+        >
+          <IndexPage />
+        </SuiProvider>
+      </ThemeProvider>
+    </StrictMode>
   )
 }
