--- packages/frontend/src/components/Main.tsx
@@ -1,36 +1,0 @@
-import '@mysten/dapp-kit/dist/index.css'
-import '@radix-ui/themes/styles.css'
-import '@suiware/kit/main.css'
-import SuiProvider from '@suiware/kit/SuiProvider'
-import { FC, StrictMode } from 'react'
-import App from '~~/components/App'
-import { APP_NAME } from '~~/config/main'
-import { getThemeSettings } from '~~/helpers/theme'
-import useNetworkConfig from '~~/hooks/useNetworkConfig'
-import ThemeProvider from '~~/providers/ThemeProvider'
-import '~~/styles/index.css'
-import { ENetwork } from '~~/types/ENetwork'
-
-const themeSettings = getThemeSettings()
-
-const Main: FC = () => {
-  const { networkConfig } = useNetworkConfig()
-
-  return (
-    <StrictMode>
-      <ThemeProvider>
-        <SuiProvider
-          customNetworkConfig={networkConfig}
-          defaultNetwork={ENetwork.LOCALNET}
-          walletAutoConnect={false}
-          walletStashedName={APP_NAME}
-          themeSettings={themeSettings}
-        >
-          <App />
-        </SuiProvider>
-      </ThemeProvider>
-    </StrictMode>
-  )
-}
-
-export default Main
