+++ packages/frontend/src/main.tsx
@@ -1,4 +1,4 @@
 import ReactDOM from 'react-dom/client'
-import Main from './components/Main'
+import App from './components/App'
 
-ReactDOM.createRoot(document.getElementById('root')!).render(<Main />)
+ReactDOM.createRoot(document.getElementById('root')!).render(<App />)
